refactor(signal-server): extract embeddable run(listener); configurable bind; logs to stderr
- Move the accept loop + connection handling from main into
  `webrtc_signal_server::run(listener)` so it can be embedded in-process
  (CLI `simulate` + e2e tests) on an ephemeral port, not just run as the
  standalone binary.
- main: bind configurable via MPC_SIGNAL_BIND (default 0.0.0.0:9000).
- All operational logs `println!` → `eprintln!` so embedders keep stdout
  clean for machine-readable output.
- Cloudflare worker route → panda.qzz.io.
